requirements.txt
Flask
google-api-python-client
Pillow
oauth2client
